Rewrote load-generator: - Using a template now instead of a function that returns a function as a string - Renamed and restructured some files - Changed README.md
