+++ load-generator/config/definitions.proto
@@ -1,0 +1,1 @@
+// Paste your proto definitions here
