--- load-generator/config/service.proto
@@ -1,1 +1,0 @@
-// Paste your proto definitions here
